int8_t for error returning and status

+++ include/basic/account_connections.h
@@ -17,14 +17,14 @@ void account_connections_set(struct account_connections *, const struct account_
 void account_connections_copy(struct account_connections *, const struct account_connections *);
 
 void account_connections_clear(struct account_connections *);
-int account_connections_cmp(const struct account_connections *, const struct account_connections *);
+int8_t account_connections_cmp(const struct account_connections *, const struct account_connections *);
 
 // Data Methods
 void account_connections_data_init(struct account_connections *);
 void account_connections_data_free(struct account_connections *);
 
 // TLV Methods
-int account_connections_set_tlv(struct account_connections *, const struct string_st *);
+int8_t account_connections_set_tlv(struct account_connections *, const struct string_st *);
 void account_connections_get_tlv(const struct account_connections *, struct string_st *);
 
 // Attrib Methods
