+++ include/operation/op_function.h
@@ -23,12 +23,12 @@ void op_function_mark(struct op_function *);
 void op_function_unmark(struct op_function *);
 
 void op_function_clear(struct op_function *);
-int op_function_cmp(const struct op_function *, const struct op_function *);
+int8_t op_function_cmp(const struct op_function *, const struct op_function *);
 
 void op_function_define(struct op_function *, size_t, struct parser_st *);
 
 // TLV Methods
-int op_function_set_tlv(struct op_function *, const struct string_st *);
+int8_t op_function_set_tlv(struct op_function *, const struct string_st *);
 void op_function_get_tlv(const struct op_function *, struct string_st *);
 
 // Convert Methods
